requirements.txt
aiogram
python-dotenv
openai
gtts
apscheduler
pydub
